small formatting improvements, moved hclGen and helper functions to their own file
