+++ internal/cli/init.go
@@ -1,15 +1,11 @@
 package cli
 
 import (
-	"embed"
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"net/url"
 	"os"
 	"path/filepath"
-	"sort"
-	"strconv"
 	"strings"
 	"time"
 
@@ -21,15 +17,13 @@ import (
 
 	"github.com/hashicorp/waypoint-plugin-sdk/terminal"
 
-	embedJson "github.com/hashicorp/waypoint/embedJson"
+	hclpkg "github.com/hashicorp/waypoint/hclgen"
 	"github.com/hashicorp/waypoint/internal/cli/datagen"
 	clientpkg "github.com/hashicorp/waypoint/internal/client"
 	"github.com/hashicorp/waypoint/internal/clierrors"
 	configpkg "github.com/hashicorp/waypoint/internal/config"
 	"github.com/hashicorp/waypoint/internal/datasource"
 	"github.com/hashicorp/waypoint/internal/pkg/flag"
-	"github.com/hashicorp/waypoint/internal/plugin"
-	fmtpkg "github.com/hashicorp/waypoint/pkg/config"
 	pb "github.com/hashicorp/waypoint/pkg/server/gen"
 	serverptypes "github.com/hashicorp/waypoint/pkg/server/ptypes"
 )
@@ -175,18 +169,14 @@ func (c *InitCommand) Run(args []string) int {
 				clierrors.Humanize(err),
 				terminal.WithErrorStyle(),
 			)
+			return 1
 		} else if strings.ToLower(proceed) == "yes" || strings.ToLower(proceed) == "y" {
 			c.ui.Output("Starting interactive .hcl generator.\n")
-			if !c.hclGen() {
-				return 1
-			}
-		} else if strings.ToLower(proceed) == "no" {
-			c.ui.Output("Generating template file.\n")
-			if !c.initNew() {
+			if !hclpkg.HclGen(c.ui) {
 				return 1
 			}
 		} else {
-			c.ui.Output("Input did not match any option, generating template file\n")
+			c.ui.Output("Generating template file\n")
 			if !c.initNew() {
 				return 1
 			}
@@ -262,667 +252,6 @@ validate the configuration and initialize your project.
 	)
 
 	return true
-}
-
-type PlugDocs struct {
-	PlugDocs []struct {
-		PlugSubDocs []struct {
-			Field    string `json:"Field"`
-			Type     string `json:"Type"`
-			Synopsis string `json:"Synopsis"`
-			Optional bool   `json:"Optional"`
-			Category bool   `json:"Category"`
-			EnvVar   string `json:"EnvVar"`
-		} `json:"SubFields"`
-		Field    string `json:"Field"`
-		Type     string `json:"Type"`
-		Synopsis string `json:"Synopsis"`
-		Optional bool   `json:"Optional"`
-		Category bool   `json:"Category"`
-		EnvVar   string `json:"EnvVar"`
-	} `json:"requiredFields"`
-	Name        string `json:"name"`
-	Description string `json:"description"`
-	Type        string `json:"type"`
-}
-
-type fieldInfo struct {
-	contents string
-	isParent bool
-	children map[string]string
-}
-
-func (c *InitCommand) hclGen() bool {
-	brackets := 0
-	hclFile, err := os.Create("waypoint.hcl")
-	if err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	}
-	c.ui.Output("Initial waypoint.hcl created!", terminal.WithStyle(terminal.SuccessBoldStyle))
-	c.ui.Output("Type \"exit\" at any point to exit the generator")
-	c.ui.Output("Name your project", terminal.WithHeaderStyle())
-	projName, err, close := c.getName("project")
-	if err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	} else if close {
-		c.exitSafe(hclFile, brackets)
-		return false
-	}
-	hclFile.Write([]byte(fmt.Sprintf("project = \"%s\"\n", projName)))
-	c.ui.Output("Name your app", terminal.WithHeaderStyle())
-	appName, err, close := c.getName("app")
-	if err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	} else if close {
-		c.exitSafe(hclFile, brackets)
-		return false
-	}
-	hclFile.Write([]byte(fmt.Sprintf("app \"%s\" {\n", appName)))
-	brackets++
-
-	var pluginNames []string
-	for pluginName := range plugin.Builtins {
-		pluginNames = append(pluginNames, pluginName)
-	}
-	var fList []string
-	dirList, _ := embedJson.Files.ReadDir("gen")
-	for _, dirE := range dirList {
-		fList = append(fList, dirE.Name())
-	}
-
-	c.ui.Output(
-		"Choose build, registry, deployment platform, and releaser plugins",
-		terminal.WithHeaderStyle(),
-	)
-
-	c.ui.Output("Configure builder", terminal.WithHeaderStyle())
-	// Select a builder
-	plug, err, close := c.selectPlugin(1, fList, embedJson.Files)
-	if err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	} else if close {
-		c.exitSafe(hclFile, brackets)
-		return false
-	}
-	hclFile.Write([]byte(fmt.Sprintf(c.genIndent(brackets) + "build {\n")))
-	brackets++
-	if plug.Name != "" {
-		hclFile.Write([]byte(fmt.Sprintf(c.genIndent(brackets)+"use \"%s\" {\n", plug.Name)))
-		brackets++
-		fieldMap, err, close := c.populatePlugins(plug)
-		if err != nil {
-			c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-			return false
-		} else if close {
-			c.exitSafe(hclFile, brackets)
-			return false
-		}
-		err = c.writeFields(hclFile, fieldMap, brackets)
-		if err != nil {
-			c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-			return false
-		}
-		c.ui.Output(
-			"Step complete: builder configuration",
-			terminal.WithSuccessStyle(),
-		)
-	}
-
-	// Here we want to close a bracket so that the registry does not appear in the "use" stanza
-	err = c.closeBrackets(hclFile, 1, brackets)
-	if err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	}
-	brackets--
-
-	c.ui.Output("Configure registry", terminal.WithHeaderStyle())
-	// Select a registry
-	plug, err, close = c.selectPlugin(4, fList, embedJson.Files)
-	if err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	} else if close {
-		c.exitSafe(hclFile, brackets)
-		return false
-	}
-	// A registry stanza will only appear in the file if one is chosen
-	if plug.Name != "" {
-		hclFile.Write([]byte(fmt.Sprintf(c.genIndent(brackets) + "registry {\n")))
-		brackets++
-		hclFile.Write([]byte(fmt.Sprintf(c.genIndent(brackets)+"use \"%s\" {\n", plug.Name)))
-		brackets++
-		fieldMap, err, close := c.populatePlugins(plug)
-		if err != nil {
-			c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-			return false
-		} else if close {
-			c.exitSafe(hclFile, brackets)
-			return false
-		}
-		err = c.writeFields(hclFile, fieldMap, brackets)
-		if err != nil {
-			c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		}
-		c.ui.Output(
-			"Step complete: registry configuration",
-			terminal.WithSuccessStyle(),
-		)
-	}
-
-	// After the registry stanza we want to close the brackets on the build
-	// and registry (if it exists) stanzas
-	err = c.closeBrackets(hclFile, brackets-1, brackets)
-	if err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	}
-	brackets = 1
-
-	c.ui.Output("Configure deployment platform", terminal.WithHeaderStyle())
-	// Select a deployer
-	plug, err, close = c.selectPlugin(2, fList, embedJson.Files)
-	if err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	} else if close {
-		c.exitSafe(hclFile, brackets)
-		return false
-	}
-
-	// A deployer stanza will only appear in the file if one is chosen
-	if plug.Name != "" {
-		hclFile.Write([]byte(fmt.Sprintf(c.genIndent(brackets) + "deploy {\n")))
-		brackets++
-		hclFile.Write([]byte(fmt.Sprintf(c.genIndent(brackets)+"use \"%s\" {\n", plug.Name)))
-		brackets++
-		fieldMap, err, close := c.populatePlugins(plug)
-		if err != nil {
-			c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-			return false
-		} else if close {
-			c.exitSafe(hclFile, brackets)
-			return false
-		}
-		err = c.writeFields(hclFile, fieldMap, brackets)
-		if err != nil {
-			c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		}
-		c.ui.Output(
-			"Step complete: deployment platform configuration",
-			terminal.WithSuccessStyle(),
-		)
-	}
-	// After the deployer stanza we want to close the brackets on the deployer stanza
-	err = c.closeBrackets(hclFile, brackets-1, brackets)
-	if err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	}
-	brackets = 1
-
-	c.ui.Output("Configure releaser", terminal.WithHeaderStyle())
-	// Select a releaser
-	plug, err, close = c.selectPlugin(3, fList, embedJson.Files)
-	if err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	} else if close {
-		c.exitSafe(hclFile, brackets)
-		return false
-	}
-
-	// A releaser stanza will only appear in the file if one is chosen
-	if plug.Name != "" {
-		hclFile.Write([]byte(fmt.Sprintf(c.genIndent(brackets) + "release {\n")))
-		brackets++
-		hclFile.Write([]byte(fmt.Sprintf(c.genIndent(brackets)+"use \"%s\" {\n", plug.Name)))
-		brackets++
-		fieldMap, err, close := c.populatePlugins(plug)
-		if err != nil {
-			c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-			return false
-		} else if close {
-			c.exitSafe(hclFile, brackets)
-			return false
-		}
-		err = c.writeFields(hclFile, fieldMap, brackets)
-		if err != nil {
-			c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		}
-		c.ui.Output("Step complete: releaser configuration", terminal.WithSuccessStyle())
-	}
-	// After the releaser stanza we want to close all the brackets
-	err = c.closeBrackets(hclFile, brackets, brackets)
-	if err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	}
-	hclFile.Close()
-	c.ui.Output("\nAll plugin configuration complete", terminal.WithSuccessStyle())
-	c.ui.Output("\nwaypoint.hcl saved!", terminal.WithStyle(terminal.SuccessBoldStyle))
-	c.ui.Output(
-		"\nIf you skipped any steps, open your waypoint.hcl file to add missing plugins or fields before continuing. (See https://www.waypointproject.io/plugins)",
-	)
-	c.ui.Output("Otherwise, run \"waypoint init\" again to start using Waypoint!\n")
-	c.ui.Output("Now attempting to format the HCL file:\n")
-	b, err := ioutil.ReadFile("waypoint.hcl")
-	if err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	}
-	out, err := fmtpkg.Format(b, "waypoint.hcl")
-	if err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	}
-	if err := ioutil.WriteFile("waypoint.hcl", out, 0644); err != nil {
-		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
-		return false
-	}
-	c.ui.Output("\nFormatting successful!", terminal.WithSuccessStyle())
-	return true
-}
-
-func (c *InitCommand) writeFields(file *os.File, fieldMap map[string]fieldInfo, brackets int) error {
-	for key, elem := range fieldMap {
-		if elem.isParent {
-			file.Write([]byte(fmt.Sprintf(c.genIndent(brackets)+"%s {\n", key)))
-			brackets++
-			for name, cont := range elem.children {
-				file.Write([]byte(fmt.Sprintf(c.genIndent(brackets)+"%s = \"%s\"\n", name, cont)))
-			}
-			err := c.closeBrackets(file, 1, brackets)
-			if err != nil {
-				return err
-			}
-			brackets--
-		} else {
-			file.Write([]byte(fmt.Sprintf(c.genIndent(brackets)+"%s = \"%s\"\n", key, elem.contents)))
-		}
-	}
-	return nil
-}
-
-func (c *InitCommand) exitSafe(file *os.File, outstanding int) error {
-	c.closeBrackets(file, outstanding, outstanding)
-	file.Close()
-	c.ui.Output("Generator exited. Any information you added before exiting has been included in your waypoint.hcl file. Edit this file manually before using Waypoint.")
-	return nil
-}
-
-func (c *InitCommand) closeBrackets(file *os.File, toClose int, outstanding int) error {
-	extra := outstanding - toClose
-	toPrint := ""
-	for i := toClose; i > 0; i-- {
-		for k := extra + i; k > 1; k-- {
-			toPrint = toPrint + "    "
-		}
-		toPrint = toPrint + "}\n"
-		file.Write([]byte(fmt.Sprintf(toPrint)))
-		toPrint = ""
-	}
-	return nil
-}
-
-func (c *InitCommand) genIndent(outstanding int) string {
-	spaces := ""
-	for i := outstanding; i > 0; i-- {
-		spaces = spaces + "    "
-	}
-	return spaces
-}
-
-func (c *InitCommand) populatePlugins(plug PlugDocs) (map[string]fieldInfo, error, bool) {
-	m := make(map[string]fieldInfo)
-	fCount := 0
-	for _, f := range plug.PlugDocs {
-		if f.Category {
-			for _, sf := range f.PlugSubDocs {
-				if !sf.Optional {
-					fCount++
-				}
-			}
-		} else {
-			fCount++
-		}
-	}
-	if plug.PlugDocs == nil || fCount == 0 {
-		c.ui.Output(
-			"There are no required fields for this %s plugin, but there may be optional fields you can add to your .hcl file later. See the Waypoint plugin documentation for more information.",
-			plug.Type,
-		)
-	} else {
-		if fCount == 1 {
-			c.ui.Output(
-				"Please complete the following %d required field for %s, or hit \"return\" to skip.",
-				fCount, plug.Name,
-				terminal.WithHeaderStyle(),
-			)
-		} else {
-			c.ui.Output(
-				"Please complete the following %d required fields for %s, or hit \"return\" to skip.",
-				fCount, plug.Name,
-				terminal.WithHeaderStyle(),
-			)
-		}
-		for _, field := range plug.PlugDocs {
-			if field.Category {
-				// Subfield handling
-				m[field.Field] = fieldInfo{isParent: true, children: make(map[string]string)}
-				for _, sfield := range field.PlugSubDocs {
-					if !sfield.Optional {
-						cont, err, close := c.populateField(sfield.Field, sfield.Type)
-						if err != nil {
-							return m, err, false
-						} else if close {
-							return m, nil, true
-						}
-						m[field.Field].children[sfield.Field] = cont
-					}
-				}
-			} else {
-				cont, err, close := c.populateField(field.Field, field.Type)
-				if err != nil {
-					return m, err, false
-				} else if close {
-					return m, nil, true
-				}
-				m[field.Field] = fieldInfo{contents: cont, isParent: false}
-			}
-		}
-	}
-	return m, nil, false
-}
-
-func (c *InitCommand) populateField(name string, fType string) (string, error, bool) {
-	getField := true
-	typeString := fType
-	if typeString == "" {
-		typeString = "No_Type_Specified"
-	}
-	for getField {
-		fieldVal, err := c.ui.Input(&terminal.Input{
-			Prompt: fmt.Sprintf("%s <%s>: ", strings.Title(name), typeString),
-			Style:  "",
-			Secret: false,
-		})
-		if err != nil {
-			c.ui.Output(
-				"Error getting input: %s",
-				clierrors.Humanize(err),
-				terminal.WithErrorStyle(),
-			)
-			return "", err, false
-		} else if strings.ToLower(fieldVal) == "exit" {
-			return "", nil, true
-		} else if strings.ToLower(fieldVal) == "" {
-			c.ui.Output(fmt.Sprintf("You have selected to skip the %s field.", name))
-			pNameConfirm, err := c.ui.Input(&terminal.Input{
-				Prompt: fmt.Sprintf("Do you really want to skip the %s field? (y/N): ", name),
-				Style:  "",
-				Secret: false,
-			})
-			if err != nil {
-				c.ui.Output(
-					"Error getting input: %s",
-					clierrors.Humanize(err),
-					terminal.WithErrorStyle(),
-				)
-				return "", err, false
-			} else if strings.ToLower(pNameConfirm) == "exit" {
-				return "", nil, true
-			} else if strings.ToLower(pNameConfirm) == "yes" || strings.ToLower(pNameConfirm) == "y" {
-				c.ui.Output("%s skipped\n", strings.Title(name), terminal.WithWarningStyle())
-				return "", nil, false
-			} else {
-				c.ui.Output("Skip cancelled\n")
-			}
-		} else {
-			// TODO: field input type checking
-			c.ui.Output("You inputted \"%s\"\n", fieldVal)
-			fieldConfirm, err := c.ui.Input(&terminal.Input{
-				Prompt: "Is this correct? (y/N): ",
-				Style:  "",
-				Secret: false,
-			})
-			if err != nil {
-				c.ui.Output(
-					"Error getting input: %s",
-					clierrors.Humanize(err),
-					terminal.WithErrorStyle(),
-				)
-				return "", err, false
-			} else if strings.ToLower(fieldConfirm) == "exit" {
-				return "", nil, true
-			} else if strings.ToLower(fieldConfirm) == "yes" || strings.ToLower(fieldConfirm) == "y" {
-				c.ui.Output("%s confirmed\n", strings.Title(name), terminal.WithSuccessStyle())
-				return fieldVal, nil, false
-			} else {
-				c.ui.Output("%s rejected\n", strings.Title(name))
-			}
-		}
-	}
-	return "", nil, true
-}
-
-// <plug> indicates the plugin that the user needs to select. 1: Builder, 2: Deployer/Platform, 3: Releaser, 4: Registry
-func (c *InitCommand) selectPlugin(plug int, fList []string, fSystem embed.FS) (PlugDocs, error, bool) {
-	var plugType string
-	var plugDocs PlugDocs
-	switch plug {
-	case 1:
-		plugType = "builder"
-	case 2:
-		plugType = "deployment platform"
-	case 3:
-		plugType = "releaser"
-	case 4:
-		plugType = "registry"
-	}
-	var plugList []string
-	for _, file := range fList {
-		if filepath.Ext(file) == ".json" {
-			switch plug {
-			case 1:
-				if strings.HasPrefix(file, "builder") {
-					plugList = append(plugList, file)
-				}
-			case 2:
-				if strings.HasPrefix(file, "platform") {
-					plugList = append(plugList, file)
-				}
-			case 3:
-				if strings.HasPrefix(file, "release") {
-					plugList = append(plugList, file)
-				}
-			case 4:
-				if strings.HasPrefix(file, "registry") {
-					plugList = append(plugList, file)
-				}
-			}
-		}
-	}
-	sort.Strings(plugList)
-	c.ui.Output(fmt.Sprintf("Select a %s: learn more at https://www.waypointproject.io/plugins. To use a %s that’s not shown here hit return, then edit the .hcl file after it’s been generated.\n",
-		plugType, plugType))
-	jMap := make(map[string]interface{})
-	var selList []string
-	var nameSelList []string
-	count := 1
-	for _, f := range plugList {
-		byteValue, err := fSystem.ReadFile(fmt.Sprintf("gen/%s", f))
-		if err != nil {
-			return plugDocs, err, false
-		}
-		json.Unmarshal(byteValue, &jMap)
-
-		// There is an assumption here that all valid plugins will have a description,
-		// we have to unmarshal all the plugins for a given stage to get an accurate name
-		// and ensure that they exist
-		if _, ok := jMap["description"]; ok {
-			c.ui.Output(fmt.Sprintf("%d: %s", count, jMap["name"]), terminal.WithInfoStyle())
-			count++
-			selList = append(selList, f)
-			nameSelList = append(nameSelList, fmt.Sprintf("%s", jMap["name"]))
-		}
-		for k := range jMap {
-			delete(jMap, k)
-		}
-	}
-	// This generates a newline after the list of plugins
-	c.ui.Output("")
-	selFileName := ""
-	getSelect := true
-	for getSelect {
-		num, err := c.ui.Input(&terminal.Input{
-			Prompt: fmt.Sprintf(
-				"Please select a plugin by typing its corresponding number or hit \"return\" to skip this step (1-%d): ",
-				count-1,
-			),
-			Style:  "",
-			Secret: false,
-		})
-		if err != nil {
-			c.ui.Output(
-				"Error getting input: %s",
-				clierrors.Humanize(err),
-				terminal.WithErrorStyle(),
-			)
-			return plugDocs, err, false
-		} else if strings.ToLower(num) == "exit" {
-			return plugDocs, nil, true
-		} else if val, err := strconv.Atoi(num); err == nil && (0 < val && val < count) {
-			c.ui.Output(fmt.Sprintf("You have selected the %s plugin.", nameSelList[val-1]))
-			pNameConfirm, err := c.ui.Input(&terminal.Input{
-				Prompt: fmt.Sprintf("Is this %s plugin correct? (y/N): ", plugType),
-				Style:  "",
-				Secret: false,
-			})
-			if err != nil {
-				c.ui.Output(
-					"Error getting input: %s",
-					clierrors.Humanize(err),
-					terminal.WithErrorStyle(),
-				)
-				return plugDocs, err, false
-			} else if strings.ToLower(pNameConfirm) == "exit" {
-				return plugDocs, nil, true
-			} else if strings.ToLower(pNameConfirm) == "yes" || strings.ToLower(pNameConfirm) == "y" {
-				c.ui.Output("%s plugin confirmed\n", strings.Title(plugType), terminal.WithSuccessStyle())
-				selFileName = selList[val-1]
-				getSelect = false
-			} else {
-				c.ui.Output("%s plugin rejected\n", strings.Title(plugType))
-			}
-		} else if num == "" {
-			c.ui.Output(fmt.Sprintf("You have selected to skip the %s stage.", plugType))
-			pNameConfirm, err := c.ui.Input(&terminal.Input{
-				Prompt: fmt.Sprintf("Do you really want to skip the %s stage? (y/N): ", plugType),
-				Style:  "",
-				Secret: false,
-			})
-			if err != nil {
-				c.ui.Output(
-					"Error getting input: %s",
-					clierrors.Humanize(err),
-					terminal.WithErrorStyle(),
-				)
-				return plugDocs, err, false
-			} else if strings.ToLower(pNameConfirm) == "exit" {
-				return plugDocs, nil, true
-			} else if strings.ToLower(pNameConfirm) == "yes" || strings.ToLower(pNameConfirm) == "y" {
-				c.ui.Output("Step complete: %s stage skipped", strings.Title(plugType), terminal.WithWarningStyle())
-				plugDocs.Name = ""
-				return plugDocs, nil, false
-			} else {
-				c.ui.Output("Skip cancelled\n")
-			}
-		} else {
-			c.ui.Output("Please select a numbered entry or type nothing to skip.\n")
-		}
-	}
-	// We again unmarshal the JSON file corresponding to the file the user has selected
-	if selFileName != "" {
-		byteValue, err := fSystem.ReadFile(fmt.Sprintf("gen/%s", selFileName))
-		if err != nil {
-			return plugDocs, err, false
-		}
-		json.Unmarshal(byteValue, &plugDocs)
-		if plugDocs.Name != "" {
-			c.ui.Output(fmt.Sprintf("You have selected the %s %s plugin.", plugDocs.Name, plugType))
-
-		} else {
-			//TODO: better error here, do we need to check again here?
-			return plugDocs, nil, true
-		}
-		return plugDocs, nil, false
-	}
-	return plugDocs, nil, false
-}
-
-// Gets either a project or app name for an HCL file, pa should be either "project" or "app"
-func (c *InitCommand) getName(pa string) (string, error, bool) {
-	if pa == "project" {
-		c.ui.Output(
-			"A project contains your app and typically maps 1:1 to a VCS repository. This name must be unique for your Waypoint server. If you're running in local mode, this must be unique to your machine.\n",
-		)
-	}
-	prompt := ""
-	if pa == "project" {
-		prompt = "Please enter a project name"
-	} else {
-		prompt = "Please enter an app name"
-	}
-	getName := true
-	name := ""
-	for getName {
-		paName, err := c.ui.Input(&terminal.Input{
-			Prompt: prompt + ": ",
-			Style:  "",
-			Secret: false,
-		})
-		if err != nil {
-			c.ui.Output(
-				"Error getting input: %s",
-				clierrors.Humanize(err),
-				terminal.WithErrorStyle(),
-			)
-			return "", err, false
-		} else if strings.ToLower(paName) == "exit" {
-			return "", nil, true
-		} else if strings.ToLower(paName) == "" {
-			c.ui.Output(prompt + ".\n")
-		} else {
-			c.ui.Output("You inputted \"%s\"\n", paName)
-			pNameConfirm, err := c.ui.Input(&terminal.Input{
-				Prompt: fmt.Sprintf("Is this %s name correct? (y/N): ", pa),
-				Style:  "",
-				Secret: false,
-			})
-			if err != nil {
-				c.ui.Output(
-					"Error getting input: %s",
-					clierrors.Humanize(err),
-					terminal.WithErrorStyle(),
-				)
-				return "", err, false
-			} else if strings.ToLower(pNameConfirm) == "exit" {
-				return "", nil, true
-			} else if strings.ToLower(pNameConfirm) == "yes" || strings.ToLower(pNameConfirm) == "y" {
-				c.ui.Output("%s name confirmed", strings.Title(pa), terminal.WithSuccessStyle())
-				name = paName
-				getName = false
-			} else {
-				c.ui.Output("%s name rejected", strings.Title(pa))
-			}
-		}
-	}
-	return name, nil, false
 }
 
 func (c *InitCommand) validateConfig() bool {
